argos_ml_project/requirements.txt
asgiref
Django
django-model-utils
djangorestframework
et-xmlfile
openpyxl
Pillow
psycopg2
pytz
sqlparse
tzdata
Unipath
wfastcgi
flask






